backend/requirements.txt
fastapi==0.117.1
uvicorn
python-jose[cryptography]
httpx==0.28.1
python-dotenv
python-multipart
